Support Linux by-path device names
Support path such as /dev/serial/by-path/pci-0000:00:11.0-usb-0:2:1.3 for
serial devices. This changes the bevaviour of the device specifier which
used to use ':' as a device delimer. However, this was not mentioned in the
docs anyway, so removed. This leaves 3 other delimiter, which should still
be enough and this is probably a seldom used feature anyway...